# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Availability": {
        "additionalProperties": false,
        "description": "When, and how much, the candidate wants to work.\n\nDistinct from `WorkAuthorization.permit_hours_cap` on purpose: this is a\npreference and feeds the `SCHEDULE_FIT` score, while a permit cap is a legal\nlimit and feeds eligibility. Collapsing the two would turn \"I would rather\nnot work Sundays\" into \"this candidate may not work Sundays\".",
        "properties": {
          "earliest_start": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Earliest Start"
          },
          "latest_end": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest End"
          },
          "max_weekly_hours": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 168.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Weekly Hours"
          },
          "min_weekly_hours": {
            "anyOf": [
              {
                "maximum": 168.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Weekly Hours"
          },
          "notice_period_days": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notice Period Days"
          },
          "weekly_slots": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/WeeklyAvailabilitySlot"
            },
            "title": "Weekly Slots",
            "type": "array"
          }
        },
        "title": "Availability",
        "type": "object"
      },
...
      "CandidateProfileDraft": {
        "additionalProperties": false,
        "description": "A candidate profile as submitted: no id, no owner, no timestamps.\n\nEvery field is the domain's own value object, so the validation a request body\ngets is exactly the validation the aggregate enforces — the language code\npattern, the availability window ordering, the permit hour cap. Nothing is\nre-specified here that `CandidateProfile` already specifies.",
        "properties": {
          "availability": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Availability"
              },
              {
                "type": "null"
              }
            ]
          },
          "base_location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ]
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "languages": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/LanguageProficiency"
            },
            "title": "Languages",
            "type": "array"
          },
          "work_authorizations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/WorkAuthorization"
            },
            "title": "Work Authorizations",
            "type": "array"
          }
        },
        "required": [
          "display_name"
        ],
        "title": "CandidateProfileDraft",
        "type": "object"
      },
      "CandidateProfileResponse": {
        "additionalProperties": false,
        "description": "A saved profile, echoed back with the two fields the draft could not carry.\n\n`evidence` and `claims` are not here either: they have no storage until Phase\n10, and a field that was always `[]` would read as \"this candidate has no\nevidence\" rather than \"this system does not keep any yet\".",
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/CandidateProfileDraft"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "updated_at",
          "profile"
        ],
        "title": "CandidateProfileResponse",
        "type": "object"
      },
...
      "GeocodingConfidence": {
        "description": "How sure the geocoder was, on a scale small enough to mean something.\n\nThree levels rather than a float: providers report confidence on scales that\ndo not compare (Nominatim's `importance` is a popularity measure, not a\nprobability), so a number here would imply an accuracy the input does not\nhave. Three buckets are enough for the one decision that reads them — §7 and\n§39's \"do not let a weaker answer replace a stronger one\" — and `rank` is what\nmakes that comparison explicit.",
        "enum": [
          "HIGH",
          "MEDIUM",
          "LOW"
        ],
        "title": "GeocodingConfidence",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LanguageProficiency": {
        "additionalProperties": false,
        "description": "A language the candidate actually has, at the level they have it.",
        "properties": {
          "language": {
            "pattern": "^[a-z]{2}$",
            "title": "Language",
            "type": "string"
          },
          "level": {
            "$ref": "#/components/schemas/LanguageLevel"
          }
        },
        "required": [
          "language",
          "level"
        ],
        "title": "LanguageProficiency",
        "type": "object"
      },
      "Location": {
        "additionalProperties": false,
        "description": "A place, at whatever precision the source actually gave us.\n\nEvery field is optional but the whole must say something: a source that\nreports only \"Yverdon-les-Bains, Suisse\" produces `raw=\"…\"` with nothing\nparsed, and a geocoding pass later fills `city`/`country`/`point`. Keeping\n`raw` is what allows that pass to be re-run and audited instead of guessed\nonce and forgotten.\n\nPhase 7 added the five fields after `raw`, and they describe the *coordinates*\nrather than the place: where the point came from, how precisely it locates\nanything, how sure the geocoder was, which geocoder it was and when. Together\nthey are what makes the enrichment pass of §21 safe to re-run — it can see\nthat a point is already better than anything it could produce and skip it —\nand what stops a city centroid being drawn as a street address (§32).\n\nNothing here is required, because a `Location` that came off a job board has\nno provenance to state beyond \"the source said so\", which is the default.",
        "properties": {
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeocodingConfidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "country": {
            "anyOf": [
              {
                "pattern": "^[A-Z]{2}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "geocoded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geocoded At"
          },
          "geocoder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geocoder"
          },
          "point": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoPoint"
              },
              {
                "type": "null"
              }
            ]
          },
          "postal_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postal Code"
          },
          "precision": {
            "$ref": "#/components/schemas/LocationPrecision",
            "default": "UNKNOWN"
          },
          "provenance": {
            "$ref": "#/components/schemas/LocationProvenance",
            "default": "SOURCE_PROVIDED"
          },
          "raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          }
        },
        "title": "Location",
        "type": "object"
      },
...
      "LocationProvenance": {
        "description": "Where a location's coordinates came from (Phase 7 §7).\n\nThe distinction has teeth: §7 forbids overwriting source-provided coordinates\nwith lower-confidence geocoder output, and a column that only stored a point\ncould not tell the two apart. `MANUAL` is a human correction and outranks\nboth — nothing automated may replace it.",
        "enum": [
          "SOURCE_PROVIDED",
          "GEOCODED",
          "MANUAL"
        ],
        "title": "LocationProvenance",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WeeklyAvailabilitySlot": {
        "additionalProperties": false,
        "description": "A recurring window the candidate can work, in whole local hours.\n\nHour granularity, and no timezone: a candidate saying \"Saturday mornings\"\nmeans it in the shop's local time, and modelling that as an instant would\ninvent precision nobody supplied.",
        "properties": {
          "end_hour": {
            "maximum": 24.0,
            "minimum": 1.0,
            "title": "End Hour",
            "type": "integer"
          },
          "start_hour": {
            "maximum": 23.0,
            "minimum": 0.0,
            "title": "Start Hour",
            "type": "integer"
          },
          "weekday": {
            "$ref": "#/components/schemas/Weekday"
          }
        },
        "required": [
          "weekday",
          "start_hour",
          "end_hour"
        ],
        "title": "WeeklyAvailabilitySlot",
        "type": "object"
      },
      "WorkAuthorization": {
        "additionalProperties": false,
        "description": "The candidate's right to work in one country.\n\n`permit_hours_cap` is what makes a whole class of eligibility deterministic:\na student permit that caps paid work at 15h/week makes a 20h/week student job\n*ineligible*, not merely a poor schedule fit. The cap itself is a legal fact\na Country Pack supplies (Phase 5); the domain only carries it.",
        "properties": {
          "country": {
            "pattern": "^[A-Z]{2}$",
            "title": "Country",
            "type": "string"
          },
          "evidence_ids": {
            "default": [],
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "permit_hours_cap": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 168.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permit Hours Cap"
          },
          "permit_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permit Label"
          },
          "status": {
            "$ref": "#/components/schemas/WorkAuthorizationStatus"
          },
          "valid_until": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          }
        },
        "required": [
          "country",
          "status"
        ],
        "title": "WorkAuthorization",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Job Search Command Center",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/me/profile": {
      "get": {
        "description": "This account's candidate profile.\n\n404 when onboarding has not saved one yet. Not an empty 200: \"no profile\" is a\ndifferent state from \"a profile with no fields filled in\", and the onboarding\nscreen has to tell them apart.",
        "operationId": "read_profile_api_v2_me_profile_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Read Profile",
        "tags": [
          "v2-me"
        ]
      },
      "put": {
        "description": "Create or replace this account's candidate profile.\n\nThe draft *is* the request model — the same value objects the aggregate holds,\nminus the id, the owner and the timestamp. There is no `user_id` field for a\nbody to set, which is the point of the draft existing at all.",
        "operationId": "save_profile_api_v2_me_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateProfileDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Profile",
        "tags": [
          "v2-me"
        ]
      }
    },
...
  }
}